More cleanup of BlockRelationshipDegreesStore interface ()
Remove `degree` and `degrees` methods that are only used in tests.
replace their usages with other `degrees` method that is used in
production too.

Also dump some thoughts about versioned degree store

+++ community/kernel-api/src/main/java/org/neo4j/storageengine/api/Degrees.java
@@ -85,7 +85,7 @@ public interface Degrees {
      * @return the outgoing degree, including loops (i.e. number of relationships on the specific node) or {@code 0} if no degree of that combination was found.
      */
     default long outgoingDegree() {
-        int degree = 0;
+        long degree = 0;
         for (int type : types()) {
             degree += outgoingDegree(type);
         }
@@ -96,7 +96,7 @@ public interface Degrees {
      * @return the incoming degree, including loops (i.e. number of relationships on the specific node) or {@code 0} if no degree of that combination was found.
      */
     default long incomingDegree() {
-        int degree = 0;
+        long degree = 0;
         for (int type : types()) {
             degree += incomingDegree(type);
         }
@@ -107,7 +107,7 @@ public interface Degrees {
      * @return the degree (i.e. number of relationships on the specific node).
      */
     default long totalDegree() {
-        int degree = 0;
+        long degree = 0;
         for (int type : types()) {
             degree += totalDegree(type);
         }
